chore: split integration tests out of unit suite
