+++ sdk/python/tests/integration/online_store/test_mongodb_online_retrieval.py
@@ -1,0 +1,165 @@
+"""MongoDB online store integration tests."""
+
+import pytest
+
+from feast.protos.feast.types.EntityKey_pb2 import (
+    EntityKey as EntityKeyProto,
+)
+from feast.protos.feast.types.Value_pb2 import Value as ValueProto
+from feast.utils import _utc_now
+from tests.universal.feature_repos.universal.feature_views import TAGS
+from tests.utils.cli_repo_creator import CliRunner, get_example_repo
+
+pytestmark = pytest.mark.integration
+pytest.importorskip("pymongo")
+
+docker_available = False
+try:
+    import docker
+    from testcontainers.mongodb import MongoDbContainer
+
+    try:
+        client = docker.from_env()
+        client.ping()
+        docker_available = True
+    except Exception:
+        pass
+except ImportError:
+    pass
+
+_requires_docker = pytest.mark.skipif(
+    not docker_available,
+    reason="Docker is not available or not running. Start Docker daemon to run these tests.",
+)
+
+
+@pytest.fixture(scope="module")
+def mongodb_container():
+    container = MongoDbContainer(
+        "mongo:latest",
+        username="test",
+        password="test",  # pragma: allowlist secret
+    ).with_exposed_ports(27017)
+    container.start()
+    yield container
+    container.stop()
+
+
+@pytest.fixture
+def mongodb_connection_string(mongodb_container):
+    exposed_port = mongodb_container.get_exposed_port(27017)
+    return f"mongodb://test:test@localhost:{exposed_port}"  # pragma: allowlist secret
+
+
+@_requires_docker
+def test_mongodb_online_features(mongodb_connection_string):
+    runner = CliRunner()
+    with runner.local_repo(
+        get_example_repo("example_feature_repo_1.py"),
+        offline_store="file",
+        online_store="mongodb",
+        teardown=False,
+    ) as store:
+        store.config.online_store.connection_string = mongodb_connection_string
+
+        driver_locations_fv = store.get_feature_view(name="driver_locations")
+        customer_profile_fv = store.get_feature_view(name="customer_profile")
+        customer_driver_combined_fv = store.get_feature_view(
+            name="customer_driver_combined"
+        )
+
+        provider = store._get_provider()
+
+        driver_key = EntityKeyProto(
+            join_keys=["driver_id"], entity_values=[ValueProto(int64_val=1)]
+        )
+        provider.online_write_batch(
+            config=store.config,
+            table=driver_locations_fv,
+            data=[
+                (
+                    driver_key,
+                    {
+                        "lat": ValueProto(double_val=0.1),
+                        "lon": ValueProto(string_val="1.0"),
+                    },
+                    _utc_now(),
+                    _utc_now(),
+                )
+            ],
+            progress=None,
+        )
+
+        customer_key = EntityKeyProto(
+            join_keys=["customer_id"], entity_values=[ValueProto(string_val="5")]
+        )
+        provider.online_write_batch(
+            config=store.config,
+            table=customer_profile_fv,
+            data=[
+                (
+                    customer_key,
+                    {
+                        "avg_orders_day": ValueProto(float_val=1.0),
+                        "name": ValueProto(string_val="John"),
+                        "age": ValueProto(int64_val=3),
+                    },
+                    _utc_now(),
+                    _utc_now(),
+                )
+            ],
+            progress=None,
+        )
+
+        customer_key = EntityKeyProto(
+            join_keys=["customer_id", "driver_id"],
+            entity_values=[ValueProto(string_val="5"), ValueProto(int64_val=1)],
+        )
+        provider.online_write_batch(
+            config=store.config,
+            table=customer_driver_combined_fv,
+            data=[
+                (
+                    customer_key,
+                    {"trips": ValueProto(int64_val=7)},
+                    _utc_now(),
+                    _utc_now(),
+                )
+            ],
+            progress=None,
+        )
+
+        assert len(store.list_entities()) == 3
+        assert len(store.list_entities(tags=TAGS)) == 2
+
+        result = store.get_online_features(
+            features=[
+                "driver_locations:lon",
+                "customer_profile:avg_orders_day",
+                "customer_profile:name",
+                "customer_driver_combined:trips",
+            ],
+            entity_rows=[
+                {"driver_id": 1, "customer_id": "5"},
+                {"driver_id": 1, "customer_id": 5},
+            ],
+            full_feature_names=False,
+        ).to_dict()
+
+        assert "lon" in result
+        assert "avg_orders_day" in result
+        assert "name" in result
+        assert result["driver_id"] == [1, 1]
+        assert result["customer_id"] == ["5", "5"]
+        assert result["lon"] == ["1.0", "1.0"]
+        assert result["avg_orders_day"] == [1.0, 1.0]
+        assert result["name"] == ["John", "John"]
+        assert result["trips"] == [7, 7]
+
+        result = store.get_online_features(
+            features=["customer_driver_combined:trips"],
+            entity_rows=[{"driver_id": 0, "customer_id": 0}],
+            full_feature_names=False,
+        ).to_dict()
+
+        assert result["trips"] == [None]
