+++ sdk/python/tests/unit/permissions/test_groups_namespaces_auth.py
